rename 'shared' directory to 'ui-components' and move StyledHost to it
