+++ packages/webcomponents/src/components/billing-form/billing-form.tsx
@@ -3,7 +3,7 @@ import BillingFormSchema, { BillingFormFields } from './billing-form-schema';
 import StateOptions from '../../utils/state-options';
 import { FormController } from '../form/form';
 import { numberOnlyHandler } from '../form/utils';
-import StyledHost from '../../utils/styled-host/styled-host';
+import { StyledHost } from '../../ui-components';
 
 /**
  * @exportedPart label: Label for inputs
