+++ packages/webcomponents/src/ui-components/index.ts
@@ -1,0 +1,2 @@
+export * from './button';
+export * from './styled-host/styled-host';
