+++ packages/webcomponents/src/ui-components/styled-host/styled-host.tsx
@@ -2,7 +2,7 @@ import { FunctionalComponent, HostAttributes } from "@stencil/core/internal";
 import { Host, h } from '@stencil/core';
 import modifiedBootstrap from './modified-bootstrap.css';
 
-const StyledHost: FunctionalComponent<HostAttributes> = (props, children) => {
+export const StyledHost: FunctionalComponent<HostAttributes> = (props, children) => {
   return (
     <Host {...props} class="body" style={{ display: 'block' }}>
       <style>{modifiedBootstrap}</style>
@@ -10,5 +10,3 @@ const StyledHost: FunctionalComponent<HostAttributes> = (props, children) => {
     </Host>
   );
 };
-
-export default StyledHost;
